change zip to tar.gz, version 0.6

+++ uninstall.sh
@@ -1,4 +1,10 @@
 #! /bin/bash
 
+# check root
+if [ "$EUID" -ne 0 ]; then
+    echo "Please run as root"
+    exit
+fi
+
 rm -rf /opt/libcbench
 rm /usr/local/bin/libcbench /usr/local/bin/libcset
